requirements.txt
GitPython==3.1.31

